implemented userCategory table

+++ Shop/BussinesLogic/Model/TransactionViewModel.cs
@@ -15,6 +15,6 @@ namespace ShopData.DTO__BusinessModels_
 
         public DateTime Date { get; set; }
 
-        public Single? Amount { get; set; }
+        public double? Amount { get; set; }
     }
 }
